framer_blog_converter/requirements.txt
lxml>=4.9.0
tqdm>=4.64.0
python-dateutil>=2.8.2
beautifulsoup4>=4.11.0
click>=8.1.0
pydantic>=2.0.0
